feat: Supply list of publish topics (#4)
We want to forward both to internal and da eventbus for the time being, to support the FCL -> FCL journey. This is part of a refactor to support a list of publish topics, which should support further flexibility in future even if we drop back down to one (da eventbus) in the near term.